refactor(dfx): move `ask_for_consent` to `dfx-core`
