+++ src/dfx-core/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod canister;
+pub mod cli;
 pub mod config;
 pub mod error;
 pub mod foundation;
